refactor: update /dashboard/expanse
