--- src/app/dashboard/expanse/receipt/page.tsx
@@ -1,3 +1,0 @@
-export default function ReceiptPage() {
-  return <div></div>;
-}
